RegisterVC afgewerkt
- klikken op buttons toont welke kleur je hebt gekozen
- startknop verschijnt pas na kiezen button
- buttons zijn niet meer geplet
- navBar in orde

+++ GROEP15_MAIV_ENROUTE/RecordingViewController.m
@@ -162,7 +162,7 @@
             [[NSUserDefaults standardUserDefaults] synchronize];
             NSLog(@"dict after delete %@", self.dictDeleteSounds);
             
-            NSLog(@"remove button int = %i", self.removeButtonInteger);
+            NSLog(@"remove button int = %li", (long)self.removeButtonInteger);
         }
     }];
 }
